+++ changelog.md
@@ -11,6 +11,7 @@
 + Add the support of reveal.js and no more only remark.js
 + Code review (refactoring, new classes, reduce number of PHPMD warnings, ...)
 + Add a timeline task (?task=timeline) (based on [https://github.com/Albejr](https://github.com/Albejr))
++ Add the support of jquery-toolbar (https://github.com/paulkinzett/toolbar)
 ```
 
 ## 1.0.6